Fix redraw with negative time
The current time can become negative if `timeOffset` is negative
and its absolute value is greater than the original current time.
The canvas is not updated because there is no active event at
the negative time. So we clear the canvas if there is neither
an active event nor the last displayed event.